rest/requirements.txt
apispec==6.0.2
astroid==2.13.3
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dill==0.3.6
Flask==2.2.2
flask-apispec==0.11.4
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
gosnomer==0.0.9
greenlet==2.0.2
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
packaging==23.0
passlib==1.7.4
pip==22.3.1
platformdirs==2.6.2
psycopg2==2.9.5
PyJWT==2.6.0
pylint==2.15.10
python-dotenv==0.21.1
setuptools==65.5.1
SQLAlchemy==2.0.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
webargs==8.2.0
Werkzeug==2.2.2
wheel==0.38.4
wrapt==1.14.1




